Split off Frame from constellation.rs
